\name{download}
\alias{download}
\title{Download data from the TCGA}
\usage{
  download(cancer, array)
}
\description{
  Downloads all the data related to a cancer type and an
  array as specified in init.
}
\details{
  Takes considerable time and memory to complete. It will
  generate a .txt file for each TCGA folder.
}

