\name{filterBarcodes}
\alias{filterBarcodes}
\title{Filter TCGA data by paired barcodes}
\usage{
  filterBarcodes(cancer, array)
}
\description{
  Filters cancer samples by paired barcodes and prepares
  the table to be copied to PostgreSQL.
}
\details{
  Uses the files generated by download and pairBarcodes. It
  will generate and save three data frames: one containing
  information for every probe, one containing the sample
  values for each probe, and one containing the sample
  barcodes (reference).
}

