\name{printPostgreSQLCode}
\alias{printPostgreSQLCode}
\title{Print code for PostgreSQL}
\usage{
  printPostgreSQLCode(cancer, array)
}
\description{
  Print the code to create the table in the PostgreSQL
  database (needs the file from pairBarcodes).
}
\details{
  The print will be displayed on the console.
}

