\name{printPostgreSQLCodeprobeinfo}
\alias{printPostgreSQLCodeprobeinfo}
\title{Print code for PostgreSQL for probeinfo tables}
\usage{
  printPostgreSQLCodeprobeinfo(cancer, array)
}
\description{
  Print the code to create the probeinfo table in the
  PostgreSQL database (needs the dataframe file from
  function download). It also outputs the code in
  writePath/*createTable.txt.
}

