\name{printPostgreSQLCode}
\alias{printPostgreSQLCode}
\title{Print code for PostgreSQL}
\usage{
  printPostgreSQLCode(cancer, array)
}
\description{
  Print the code to create the table in the PostgreSQL
  database (needs the sampleinfo file from filterBarcodes).
  It also outputs the code in writePath/*createTable.txt.
}

